Fixed more warnings and clarified BaseOperationRecorder.open_file()
Details:
- Docs: Clarified that the return type of `BaseOperationRecorder.open_file()`
  is a file-like object and that the caller is responsible for closing that
  file.
- In `cim_obj._makequalifiers()`, replaced the sorting of qualifiers
  that was previously done by sorting a list of CIMQualifier objects,
  by sorting their keys.
- In `test_cim_operations.TestCreateConnection.test_add_operation_recorders()`
  added closing the output file to avoid a ResourceWarning. In addition,
  removed the generated file (blah.yaml) again.
- In `test_cim_xml.validate_xml()`, added closing of the piped stdout and stderr
  streams to avoid a ResourceWarning.
- In the base class `test_mof_compiler.MOFTest`, added a `tearDown()` method
  that closes the log file again, to avoid a ResourceWarning.
- In `test_mof_compiler.setUpModule()`, printed the empty line only if any
  of the extra actions like downloading the CIM schema is performed. This
  cleans up the dots issued by y.test for test_mof_compiler.py for the
  normal cases where such extra actions are not performed.
- In `test_mof_compiler.TestFullSchema.test_mof_schema_roundtrip()`,
  changed the sorting of qualifiers that was previously done by sorting a list
  of CIMQualifierDeclaration objects, by sorting their keys.
  While being there, reorganized some of the assertions to be more logical.
- In `validate.validate_xml()`, added closing of the stdout stream, to
  avoid a ResourceWarning.
- In `test_tupleparse.ParseCIMInstanceName.test_all()`, changed a keybinding
  from int to Uint8 to avoid a DeprecationWarning.

+++ testsuite/test_tupleparse.py
@@ -73,7 +73,7 @@ class ParseCIMInstanceName(TupleTest):
 
         self._run_single(
             CIMInstanceName('CIM_Foo', {'Name': 'Foo',
-                                        'Number': 42,
+                                        'Number': Uint8(42),
                                         'Boolean': False,
                                         'Ref': CIMInstanceName('CIM_Bar')}))
